Moved team balancing to round start in stead of round end. Fixes issue with maps or plugins inspecting player teams in round end event; being unable to detect which players that are zombies when the round has ended. Thanks to "Neuro Toxin".
Possible side effect: After round end, but before restart, zombies may not be able to pick up weapons to use in the next round. This might be ok, it's just dumb zombies. But it should be possible to fix this in the weapon restrict module if there's a need for it.